[plugin] decouple unit tests from monaco and browser dependencies
