+++ packages/plugin-ext/src/main/browser/menus/menus-contribution-handler.spec.ts
@@ -36,69 +36,70 @@ import { ResourceContextKey } from '@theia/core/lib/browser/resource-context-key
 
 disableJSDOM();
 
-let testContainer: Container;
-let handler: MenusContributionPointHandler;
-
-let notificationWarnSpy: sinon.SinonSpy;
-let registerMenuSpy: sinon.SinonSpy;
-let registerCmdHandlerSpy: sinon.SinonSpy;
-let loggerWarnSpy: sinon.SinonSpy;
-
-const testCommandId = 'core.about';
-
-before(() => {
-    testContainer = new Container();
-
-    const module = new ContainerModule((bind, unbind, isBound, rebind) => {
-        bind(ILogger).to(MockLogger).inSingletonScope();
-        bind(MessageClient).toSelf().inSingletonScope();
-        bind(MessageService).toSelf().inSingletonScope();
-        bind(MenuModelRegistry).toConstantValue(new MockMenuModelRegistry());
-        bindContributionProvider(bind, CommandContribution);
-        bind(CommandRegistry).toSelf().inSingletonScope();
-        bind(ContextKeyService).toSelf().inSingletonScope();
-        bind(MenusContributionPointHandler).toSelf();
-        // tslint:disable-next-line:no-any mock QuickCommandService
-        bind(QuickCommandService).toConstantValue({} as any);
-        // tslint:disable-next-line:no-any mock TabBarToolbarRegistry
-        bind(TabBarToolbarRegistry).toConstantValue({} as any);
-        // tslint:disable-next-line:no-any mock PluginSharedStyle
-        bind(PluginSharedStyle).toConstantValue({} as any);
-        bind(SelectionService).toSelf().inSingletonScope();
-        // tslint:disable-next-line:no-any mock ScmService
-        bind(ScmService).toConstantValue({} as any);
-        // tslint:disable-next-line:no-any mock ScmService
-        bind(ResourceContextKey).toConstantValue({} as any);
-    });
-
-    testContainer.load(module);
-});
-
-beforeEach(() => {
-    handler = testContainer.get(MenusContributionPointHandler);
-
-    const logger = testContainer.get<ILogger>(ILogger);
-    loggerWarnSpy = sinon.spy(logger, 'warn');
-
-    const messageService = testContainer.get(MessageService);
-    notificationWarnSpy = sinon.spy(messageService, 'warn');
-
-    const menuRegistry = testContainer.get(MenuModelRegistry);
-    registerMenuSpy = sinon.spy(menuRegistry, 'registerMenuAction');
-
-    const commandRegistry = testContainer.get(CommandRegistry);
-    registerCmdHandlerSpy = sinon.spy(commandRegistry, 'registerHandler');
-});
-
-afterEach(function () {
-    notificationWarnSpy.restore();
-    registerMenuSpy.restore();
-    registerCmdHandlerSpy.restore();
-    loggerWarnSpy.restore();
-});
-
 // TODO: enable tests once the https://github.com/theia-ide/theia/issues/3344 is fixed
 describe.skip('MenusContributionHandler', () => {
+
+    let testContainer: Container;
+    let handler: MenusContributionPointHandler;
+
+    let notificationWarnSpy: sinon.SinonSpy;
+    let registerMenuSpy: sinon.SinonSpy;
+    let registerCmdHandlerSpy: sinon.SinonSpy;
+    let loggerWarnSpy: sinon.SinonSpy;
+
+    const testCommandId = 'core.about';
+
+    before(() => {
+        testContainer = new Container();
+
+        const module = new ContainerModule((bind, unbind, isBound, rebind) => {
+            bind(ILogger).to(MockLogger).inSingletonScope();
+            bind(MessageClient).toSelf().inSingletonScope();
+            bind(MessageService).toSelf().inSingletonScope();
+            bind(MenuModelRegistry).toConstantValue(new MockMenuModelRegistry());
+            bindContributionProvider(bind, CommandContribution);
+            bind(CommandRegistry).toSelf().inSingletonScope();
+            bind(ContextKeyService).toSelf().inSingletonScope();
+            bind(MenusContributionPointHandler).toSelf();
+            // tslint:disable-next-line:no-any mock QuickCommandService
+            bind(QuickCommandService).toConstantValue({} as any);
+            // tslint:disable-next-line:no-any mock TabBarToolbarRegistry
+            bind(TabBarToolbarRegistry).toConstantValue({} as any);
+            // tslint:disable-next-line:no-any mock PluginSharedStyle
+            bind(PluginSharedStyle).toConstantValue({} as any);
+            bind(SelectionService).toSelf().inSingletonScope();
+            // tslint:disable-next-line:no-any mock ScmService
+            bind(ScmService).toConstantValue({} as any);
+            // tslint:disable-next-line:no-any mock ScmService
+            bind(ResourceContextKey).toConstantValue({} as any);
+        });
+
+        testContainer.load(module);
+    });
+
+    beforeEach(() => {
+        handler = testContainer.get(MenusContributionPointHandler);
+
+        const logger = testContainer.get<ILogger>(ILogger);
+        loggerWarnSpy = sinon.spy(logger, 'warn');
+
+        const messageService = testContainer.get(MessageService);
+        notificationWarnSpy = sinon.spy(messageService, 'warn');
+
+        const menuRegistry = testContainer.get(MenuModelRegistry);
+        registerMenuSpy = sinon.spy(menuRegistry, 'registerMenuAction');
+
+        const commandRegistry = testContainer.get(CommandRegistry);
+        registerCmdHandlerSpy = sinon.spy(commandRegistry, 'registerHandler');
+    });
+
+    afterEach(function () {
+        notificationWarnSpy.restore();
+        registerMenuSpy.restore();
+        registerCmdHandlerSpy.restore();
+        loggerWarnSpy.restore();
+    });
+
     describe('should register an item in the supported menus', () => {
         it('editor context menu', () => {
             handler.handle({
